Forked from tryhaskell to tryr
This project aims to provide a web-based console that will attempt to
teach the R Language for Statistical Computing.  The initial codebase of
this project is a fork of the tryhaskell.org code written by Chris Done.

Besides a change of language (Haskell->R) at the time the fork was
initiated (July 2010) the following architectural changes have been
decided on:

  * All files related to the website will be generated by the Ruby
    language and served using the Sinatra framework as opposed to using
    PHP.

  * Websockets will be used instead of AJAX to communicate between the
    web console and the R server.