Worked on connection the R terminal to the WebSocket. Problems occurred.
Basically, something screwy is going on- the responses seem to lag the
commands by 1 evaluation.  I.e.:

  > 1

Returns nothing.  Then:

  > 2+2

Returns:

  "[1] 1"

Which is the result of the first command, not the second.  Could be a
bug in the way I implemented the rServe controller or it could be due to
the fact that the WebSocket communication is a non-blocking call and
things are not synching up.