R Console is now operational!
The console callback function is now passed to the rServe controller
which stores it and then sends the command to Rserve.

When a response is recieved, the stored callback is executed on the
response- this causes the data to be printed and the console to be
unlocked.

Further things to be done on the console side:

  1. The callback function probably only needs to be registered once.
     this could be done after the console is connected and may require
     exporting an additional function from the jquery.console plugin.

  2. The command handling could be modified to show a "loading"
     animation while the console is locked and waiting for input.

  3. The rServe message reciever needs to be generalized to handle
     evaluation responses and evaluation errors- as well as additional
     functionality that may be added in the future ( plotting commands,
     edit commands, etc.).