More re-design
All the haml and scss code has been re-written from scratch with the
exception of the stuff in:

  views/stylesheets/partials/_console.scss

Which has been taken from a modified version Chris Done's tryhaskell.css
after running it through sass convert.

The web interface looks good enough to call this alpha 0.0a.  It works,
but there are many errors on both the client and server sides that are
not handled in any other way than to cause the system to crash.  The
client is not notified of this.